+++ app/src/main/java/com/example/thecoffee/order/model/Cart.kt
@@ -3,7 +3,7 @@ package com.example.thecoffee.order.model
 data class Cart (
     // item - product -> drink, topping
     // sluong
-    val totalPrice: Int?,
+    val totalPrice: Long?,
     val quantity: Int?,
     val drinkName: String?,
     val topping: List<String>?
